Move constraint tests to dedicated test_nchoosek.py and test_product.py
